rename the new sampler to sampler
